requirements.txt
alembic==1.13.1
bcrypt==4.1.3
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
decord==0.6.0
dnspython==2.6.1
email-validator==2.1.2
flask==3.0.3
Flask-Login==0.6.3
flask-mail==0.10.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security-Too==5.4.3
flask-sqlalchemy==3.1.1
Flask-User==1.0.2.2
flask-wtf==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.0
pandas==2.0.3
passlib==1.7.4
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
typing-extensions==4.12.1
tzdata==2024.1
werkzeug==3.0.3
wtforms==3.1.2
zipp==3.19.1
